+++ catalog/app/components/Preview/index.js
@@ -1,6 +1,6 @@
 export { default as Display, bind as display } from './Display'
 export { default as render } from './render'
-export { default as load, getRenderProps } from './load'
+export { default as load, getRenderProps, Load } from './load'
 export { PreviewData, PreviewError, CONTEXT } from './types'
 export { default as Header } from './Header'
 export { default as Menu } from './Menu'